requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pg8000
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
python-dotenv
alembic
email-validator
aiofiles
cryptography
